Combine :factory and :nightmode into single :skylight-android module
