--- factory/src/main/AndroidManifest.xml
@@ -1,3 +1,0 @@
-<manifest
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.drewhamilton.skylight.android.factory" />
