--- nightmode/api/nightmode.api
@@ -1,22 +1,0 @@
-public abstract class dev/drewhamilton/skylight/android/nightmode/AutoNightActivity : androidx/appcompat/app/AppCompatActivity {
-	public fun <init> ()V
-	protected fun getAutoNightDelegate ()Ldev/drewhamilton/skylight/android/nightmode/AutoNightDelegate;
-	protected fun onCreate (Landroid/os/Bundle;)V
-}
-
-public final class dev/drewhamilton/skylight/android/nightmode/AutoNightDelegate : androidx/lifecycle/DefaultLifecycleObserver {
-	public static final field Companion Ldev/drewhamilton/skylight/android/nightmode/AutoNightDelegate$Companion;
-	public fun <init> (Landroidx/appcompat/app/AppCompatDelegate;Ldev/drewhamilton/skylight/SkylightForCoordinates;)V
-	public static final fun fallback (Landroidx/appcompat/app/AppCompatDelegate;)Ldev/drewhamilton/skylight/android/nightmode/AutoNightDelegate;
-	public static final fun ofCoordinates (Landroidx/appcompat/app/AppCompatDelegate;DD)Ldev/drewhamilton/skylight/android/nightmode/AutoNightDelegate;
-	public static final fun ofTimes (Landroidx/appcompat/app/AppCompatDelegate;Ljava/time/OffsetTime;Ljava/time/OffsetTime;)Ldev/drewhamilton/skylight/android/nightmode/AutoNightDelegate;
-	public fun onStart (Landroidx/lifecycle/LifecycleOwner;)V
-	public fun onStop (Landroidx/lifecycle/LifecycleOwner;)V
-}
-
-public final class dev/drewhamilton/skylight/android/nightmode/AutoNightDelegate$Companion {
-	public final fun fallback (Landroidx/appcompat/app/AppCompatDelegate;)Ldev/drewhamilton/skylight/android/nightmode/AutoNightDelegate;
-	public final fun ofCoordinates (Landroidx/appcompat/app/AppCompatDelegate;DD)Ldev/drewhamilton/skylight/android/nightmode/AutoNightDelegate;
-	public final fun ofTimes (Landroidx/appcompat/app/AppCompatDelegate;Ljava/time/OffsetTime;Ljava/time/OffsetTime;)Ldev/drewhamilton/skylight/android/nightmode/AutoNightDelegate;
-}
-
