--- nightmode/build.gradle
@@ -1,81 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'org.jetbrains.kotlin.android'
-
-ext {
-    artifactName = 'skylight-nightmode'
-    publishedDescription = 'A delegate supporting automatic switching between night and notnight resources at dawn and dusk.'
-}
-apply from: '../publish.gradle'
-
-android {
-    compileSdkVersion projectVersions.targetSdk
-
-    defaultConfig {
-        minSdkVersion projectVersions.minSdk
-        targetSdkVersion projectVersions.targetSdk
-        versionName projectVersions.versionName
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles 'consumer-rules.pro'
-
-        // Support core library desugaring:
-        multiDexEnabled true
-    }
-
-    signingConfigs {
-        release {
-            storeFile file(keystore)
-            storePassword keystorePassword
-            keyAlias keyAlias
-            keyPassword keyPassword
-        }
-    }
-
-    buildTypes {
-        release {
-            signingConfig signingConfigs.release
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt')
-        }
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-        coreLibraryDesugaringEnabled = true
-    }
-    kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_1_8
-    }
-    lintOptions {
-        lintConfig file('../lint.xml')
-    }
-}
-
-dependencies {
-    implementation deps.kotlinStdLib
-
-    implementation deps.appCompat
-    api deps.lifecycle
-
-    api(deps.skylight) {
-        changing = versions.skylight.contains('SNAPSHOT')
-    }
-    implementation(deps.skylightCalculator) {
-        changing = versions.skylight.contains('SNAPSHOT')
-    }
-    implementation(deps.skylightFake) {
-        changing = versions.skylight.contains('SNAPSHOT')
-    }
-    implementation project(':factory')
-
-    coreLibraryDesugaring deps.coreLibraryDesugaring
-
-    testImplementation deps.junit
-    testImplementation deps.mockito
-
-    androidTestImplementation deps.junitExt
-    androidTestImplementation deps.espresso
-    androidTestImplementation project(':test')
-    androidTestImplementation deps.materialComponents
-}
