--- nightmode/src/main/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<manifest
-    package="dev.drewhamilton.skylight.android.nightmode" />
