+++ skylight-android/src/androidTest/java/dev/drewhamilton/skylight/android/TestAutoNightActivity.kt
@@ -1,0 +1,33 @@
+package dev.drewhamilton.skylight.android
+
+import android.os.Bundle
+import java.time.Instant
+import java.time.OffsetTime
+import java.time.ZoneId
+
+class TestAutoNightActivity : AutoNightActivity() {
+
+    override lateinit var autoNightDelegate: AutoNightDelegate
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        val intent = intent
+        val dawnMilli = intent.getLongExtra(KEY_DAWN_EPOCH_MILLI, -1)
+        val duskMilli = intent.getLongExtra(KEY_DUSK_EPOCH_MILLI, -1)
+        if (dawnMilli == -1L || duskMilli == -1L)
+            throw IllegalArgumentException("Dawn and dusk times must be provided")
+
+        autoNightDelegate = AutoNightDelegate.ofTimes(
+            delegate,
+            dawn = OffsetTime.ofInstant(Instant.ofEpochMilli(dawnMilli), ZoneId.systemDefault()),
+            dusk = OffsetTime.ofInstant(Instant.ofEpochMilli(duskMilli), ZoneId.systemDefault())
+        )
+
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.test_text)
+    }
+
+    companion object {
+        const val KEY_DAWN_EPOCH_MILLI = "Dawn epoch milli"
+        const val KEY_DUSK_EPOCH_MILLI = "Dusk epoch milli"
+    }
+}
