+++ skylight-android/src/debug/AndroidManifest.xml
