+++ skylight-android/src/debug/res/values-night/strings.xml
