+++ skylight-android/src/debug/res/values/strings.xml
