+++ skylight-android/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest
+    package="dev.drewhamilton.skylight.android" />
